third_party/WebKit/LayoutTests/fast/css/variables/cssom-read.html

<!doctype html>

<pre style="var-a: override me; var-a: apple; var-b: banana;"></pre>

<script>
if (window.testRunner) {
    window.testRunner.dumpAsText();
}

var pre = document.querySelector("pre");

pre.innerText += "Inline style CSS text: " + pre.style.cssText + "\n";
pre.innerText += "\n";

pre.innerText += "Count vars: " + pre.style.var.size + "\n";
pre.innerText += "\n";

pre.innerText += "Access vars directly:\n";
pre.innerText += "a: " + pre.style.var.get("a") + "\n";
pre.innerText += "b: " + pre.style.var.get("b") + "\n";
pre.innerText += "\n";

pre.innerText += "Access non existent var:\n";
pre.innerText += "z: " + pre.style.var.get("z") + "\n";
pre.innerText += "\n";

pre.innerText += "Setting c to carrot and d to dog...\n";
pre.style.var.set("c", "carrot");
pre.style.var.set("d", "dog");
pre.innerText += "\n";

pre.innerText += "Count vars: " + pre.style.var.size + "\n";
pre.innerText += "\n";

pre.innerText += "Access vars directly:\n";
pre.innerText += "a: " + pre.style.var.get("a") + "\n";
pre.innerText += "b: " + pre.style.var.get("b") + "\n";
pre.innerText += "c: " + pre.style.var.get("c") + "\n";
pre.innerText += "d: " + pre.style.var.get("d") + "\n";

</script>
